change comments from hashes to m4 dnl look also for templates default cdict usage to 'no' remove Setup file before bootstrapping.
Actually there's an interesting story here, which since you're reading the
logs I'll take the time to write up. OpenBSD 2.7's m4 is posix compliant
i believe. And to this end it limits the number of m4 diverts to 10.
Unfortunately, autoconf 2.13 uses 11 diverts because gnu's m4 has an
effectively unlimited number diverts. So some of my day was spent tracking
that bug down and bettering the GNU emulation of openbsd's m4.